Simplify the stages after review
- `find_unzip` always warns; the per-caller log level is gone.
- The bwa alignment stages report `cleanup_paths` only when the
  alignment is an intermediate that duplicate marking consumes, as the
  field's documentation already said; final outputs get none, and the
  pipelines build the `rm-*-aln` job only when there is something to
  remove.
- `TransferApplyStage`'s DNAModelApply depends on the transfer's concat
  alone (the concat already depends on the raw-VCF producer), or on
  `upstream` when there is no transfer. The hybrid pipelines now use
  the composite too, and the long-read haplotype passes each run their
  own transfer-then-apply, so a haplotype's model-apply waits only on
  its own transfer.
- `build_transfer_jobs` moves into `stages/transfer.py`, its only
  caller, and the top-level `transfer.py` is gone.
